TN-1751 followup cleanup to Consent/orgs js code
