+++ portal/static/js/src/modules/OrgTool.js
@@ -1,5 +1,7 @@
 import SYSTEM_IDENTIFIER_ENUM from "./SYSTEM_IDENTIFIER_ENUM.js";
 import tnthAjax from "./TnthAjax.js";
+import Consent from "./Consent.js";
+
 export default (function() { /*global i18next $ */
     var OrgObj = function(orgId, orgName, parentOrg) {
         this.id = orgId;
@@ -25,6 +27,7 @@ export default (function() { /*global i18next $ */
         this.setContainerElementId(orgsElementsContainerId);
         if (sessionStorage.orgsData) {
             var orgsData = JSON.parse(sessionStorage.orgsData);
+            self.orgsData = orgsData;
             self.populateOrgsList(orgsData);
             callback(orgsData);
         } else {
@@ -349,6 +352,300 @@ export default (function() { /*global i18next $ */
         if (!container.text()) {
             container.html(i18next.t("No organizations available"));
         }
+    };
+    OrgTool.prototype.populateOrgsStateSelector = function(subjectId, parentOrgsToDraw, callback) {
+        var self = this; /*global i18next */
+        callback = callback || function() {};
+        parentOrgsToDraw = parentOrgsToDraw || [];
+        var stateDict={AL: i18next.t("Alabama"),AK: i18next.t("Alaska"), AS: i18next.t("American Samoa"),AZ: i18next.t("Arizona"),AR:i18next.t("Arkansas"),CA: i18next.t("California"),CO:i18next.t("Colorado"),CT:i18next.t("Connecticut"),DE:i18next.t("Delaware"),DC:i18next.t("District Of Columbia"),FM: i18next.t("Federated States Of Micronesia"),FL:i18next.t("Florida"),GA:i18next.t("Georgia"),GU:i18next.t("Guam"),HI:i18next.t("Hawaii"),ID:i18next.t("Idaho"),IL:i18next.t("Illinois"),IN:i18next.t("Indiana"),IA:i18next.t("Iowa"),KS:i18next.t("Kansas"),KY:i18next.t("Kentucky"),LA:i18next.t("Louisiana"),ME:i18next.t("Maine"),MH:i18next.t("Marshall Islands"),MD:i18next.t("Maryland"),MA:i18next.t("Massachusetts"),MI:i18next.t("Michigan"),MN:i18next.t("Minnesota"),MS:i18next.t("Mississippi"),MO:i18next.t("Missouri"),MT:i18next.t("Montana"),NE: i18next.t("Nebraska"),NV:i18next.t("Nevada"),NH:i18next.t("New Hampshire"),NJ:i18next.t("New Jersey"),NM:i18next.t("New Mexico"),NY:i18next.t("New York"),NC:i18next.t("North Carolina"),ND:i18next.t("North Dakota"),MP:i18next.t("Northern Mariana Islands"),OH:i18next.t("Ohio"),OK:i18next.t("Oklahoma"),OR:i18next.t("Oregon"),PW:i18next.t("Palau"),PA:i18next.t("Pennsylvania"),PR:i18next.t("Puerto Rico"),RI:i18next.t("Rhode Island"),SC:i18next.t("South Carolina"),SD:i18next.t("South Dakota"),TN:i18next.t("Tennessee"),TX:i18next.t("Texas"),UT:i18next.t("Utah"),VT:i18next.t("Vermont"),VI:i18next.t("Virgin Islands"),VA:i18next.t("Virginia"),WA:i18next.t("Washington"),WV:i18next.t("West Virginia"),WI:i18next.t("Wisconsin"),WY:i18next.t("Wyoming")};
+        var getParentState = (o, states) => {
+            if (!o) {
+                return "";
+            }
+            var s = "", found = false;
+            for (var state in states) {
+                if (!found) {
+                    (states[state]).forEach(function(i) {
+                        if (String(i) === String(o)) {
+                            s = state;
+                            found = true;
+                        }
+                    });
+                }
+            }
+            return s;
+        };
+        $("#stateSelector").on("change", function() {
+            let selectedState = $(this).find("option:selected");
+            let container = $("#" + selectedState.val() + "_container");
+            let defaultPrompt = i18next.t("What is your main clinic for prostate cancer care");
+            $("#userOrgsInfo").hide();
+            if (selectedState.val() !== "") {
+                if (selectedState.val() === "none") {
+                    $(".state-container, .noOrg-container").hide();
+                    $(".clinic-prompt").text("").hide();
+                    $("#noOrgs").prop("checked", true).trigger("click");
+                    //send of ajax to update org to 0 here
+                } else {
+                    if (container.length > 0) {
+                        $(".state-container").hide();
+                        $(".clinic-prompt").text(defaultPrompt + " in " + selectedState.text() + "?").show();
+                        $(".noOrg-container").show();
+                        $("#noOrgs").prop("checked", false);
+                        container.show();
+                    } else {
+                        $(".state-container, .clinic-prompt, .noOrg-container").hide();
+                        $("#userOrgsInfo").show();
+                    }
+                }
+            } else {
+                $(".state-container, .noOrg-container").hide();
+                $(".clinic-prompt").text("").hide();
+            }
+        });
+
+        var orgsList = this.orgsList, states = {}, contentHTML = "";
+        /**** draw state select element first to gather all states - assign orgs to each state in array ***/
+        (self.orgsData).forEach(function(item) {
+            let __state = "";
+            if (!item.identifier) { return false; }
+            (item.identifier).forEach(function(region) {
+                if (String(region.system) === String(SYSTEM_IDENTIFIER_ENUM.practice_region) && region.value) {
+                    __state = (region.value).split(":")[1];
+                    __state = (region.value).split(":")[1];
+                    if (!states.hasOwnProperty(__state)) {
+                        states[__state] = [item.id];
+                        $("#userOrgs .main-state-container").prepend("<div id='" + __state + "_container' state='" + __state + "' class='state-container'></div>");
+                    } else {
+                        (states[__state]).push(item.id);
+                    }
+                    if ($("#stateSelector option[value='" + __state + "']").length === 0) {
+                        $("#stateSelector").append("<option value='" + __state + "'>" + stateDict[__state] + "</option>");
+                    }
+                    orgsList[item.id].state = __state; //assign state for each item
+                }
+            });
+        });
+
+        /*
+            * If an organization is a top level org and has child orgs, we render legend for it.  This will prevent the organization from being selected by the user.
+            * Note: a hidden input field is rendered for the organization so it can still be referenced by the child orgs if necessary.
+            */
+        var parentOrgs = $.grep(this.orgsData, function(item) {
+            return parseInt(item.id) !== 0 && !item.partOf;
+        });
+
+        parentOrgs = parentOrgs.sort(function(a, b) { //sort parent orgs so ones with children displayed first
+            var oo_1 = orgsList[a.id];
+            var oo_2 = orgsList[b.id];
+            if (oo_1 && oo_2) {
+                if (oo_1.children.length > 0 && oo_2.children.length > 0) {
+                    if (a.name < b.name) {
+                        return -1;
+                    }
+                    if (a.name > b.name) {
+                        return 1;
+                    }
+                    return 0;
+                } else if (oo_1.children.length > 0 && oo_2.children.length === 0) {
+                    return -1;
+                } else if (oo_2.children.length > 0 && oo_1.children.length === 0) {
+                    return 1;
+                } else {
+                    if (a.name < b.name) {
+                        return -1;
+                    }
+                    if (a.name > b.name) {
+                        return 1;
+                    }
+                    return 0;
+                }
+            } else {
+                return 0;
+            }
+        });
+        parentOrgs.forEach(function(item) {
+            var state = orgsList[item.id].state;
+            if ($("#" + state + "_container").length > 0) {
+                var oo = orgsList[item.id];
+                if (!(parentOrgsToDraw.indexOf(item.name) !== -1) && oo.children.length > 0) {
+                    contentHTML = `<legend orgId="${item.id}">${i18next.t(item.name)}</legend><input class="tnth-hide" type="checkbox" name="organization" parent_org="true" data-org-name="${item.name}"  id="${item.id}_org" value="${item.id}" />`;
+                } else { //also need to check for top level orgs that do not have children and render those
+                    contentHTML = `<div class="radio parent-singleton"><label><input class="clinic" type="radio" id="${item.id}_org" value="${item.id}" state="${state}" name="organization" data-parent-name="${item.name}" data-parent-id="${item.id}">${i18next.t(item.name)}</label></div>`;
+                }
+                $("#" + state + "_container").append(contentHTML);
+            }
+        });
+
+        var childOrgs = $.grep(this.orgsData, function(item) { //draw input element(s) that belongs to each state based on parent organization id
+            return parseInt(item.id) !== 0 && item.partOf;
+        });
+
+        childOrgs = childOrgs.sort(function(a, b) { //// sort child clinics in alphabetical order
+            if (a.name < b.name) {
+                return 1;
+            }
+            if (a.name > b.name) {
+                return -1;
+            }
+            return 0;
+        });
+        childOrgs.forEach(function(item) {
+            var parentId = (item.partOf.reference).split("/")[2];
+            if (parentId) {
+                if (parentOrgsToDraw.indexOf(self.getOrgName(parentId)) !== -1) {
+                    return true;
+                }
+                var parentState = getParentState(parentId, states);
+                contentHTML = `<div class="radio"><label class="indent"><input class="clinic" type="radio" id="${item.id}_org" value="${item.id}" state="${parentState}" name="organization" data-parent-name="${item.name}" data-parent-id="${parentId}">${i18next.t(item.name)}</label></div>`;
+                if ($("#" + parentState + "_container legend[orgId='" + parentId + "']").length > 0) {
+                    $("#" + parentState + "_container legend[orgId='" + parentId + "']").after(contentHTML);
+                } else {
+                    $("#" + parentState + "_container").append(contentHTML);
+                }
+            }
+        });
+        //var selectOptions = $("#stateSelector").sortOptions();
+        var selectOptions = $("#stateSelector option");
+        if (selectOptions.length > 0) {
+            var selectSortedOptions = $("#stateSelector").sortOptions();
+            if (selectSortedOptions && selectSortedOptions.length > 0) { //sorting the select options
+                $("#stateSelector").empty().append(selectSortedOptions)
+                    .append(`<option value="none">${i18next.t("Other")}</option>`)
+                    .prepend(`<option value="" selected>${i18next.t("Select")}</option>`)
+                    .val("");
+            }
+            $(".state-container, .clinic-prompt").hide();
+            setTimeout(function() { //case of pre-selected clinic, need to check if any clinic has prechecked
+                var o = $("#userOrgs input[name='organization']:checked");
+                if (o.length > 0 && parseInt(o.val()) !== 0) {
+                    o.closest(".state-container").show();
+                    $(".clinic-prompt").show();
+                }
+            }, 150);
+            $("#userOrgs input[name='organization']").each(function() {
+                if (parseInt($(this).val()) !== 0) {
+                    Consent.getDefaultModal(this);
+                }
+            });
+            self.onLoaded(subjectId, false);
+        } else { // if no states found, then need to draw the orgs UI
+            $("#userOrgs .selector-show").hide();
+            self.onLoaded(subjectId, true);
+            self.filterOrgs(self.getHereBelowOrgs());
+            self.morphPatientOrgs();
+            $(".noOrg-container, .noOrg-container *").show();
+        }
+        $("#clinics").attr("loaded", true);
+        callback();
+    };
+    OrgTool.prototype.handleOrgsEvent = function(userId, isConsentWithTopLevelOrg) {
+        var self = this;
+        $("#userOrgs input[name='organization']").each(function() {
+            $(this).attr("data-save-container-id", "userOrgs");
+            $(this).on("click", function() {
+                var parentOrg = self.getElementParentOrg(this);
+                var orgsElements = $("#userOrgs input[name='organization']").not("[id='noOrgs']");
+                if ($(this).prop("checked")) {
+                    if ($(this).attr("id") !== "noOrgs") {
+                        $("#noOrgs").prop("checked", false);
+                    } else {
+                        orgsElements.prop("checked", false);
+                    }
+                }
+                $("#userOrgs .help-block").removeClass("error-message").text("");
+                if (sessionStorage.getItem("noOrgModalViewed")) {
+                    sessionStorage.removeItem("noOrgModalViewed");
+                }
+
+                if ($(this).attr("id") !== "noOrgs" && $("#fillOrgs").attr("patient_view")) {
+                    if (tnthAjax.hasConsent(userId, parentOrg)) {
+                        self.updateOrgs(userId, $("#clinics"), true);
+                    } else {
+                        var __modal = Consent.getConsentModal(parentOrg);
+                        if (__modal && __modal.length > 0) {
+                            setTimeout(function() { __modal.modal("show"); }, 50);
+                        } else {
+                            self.updateOrgs(userId, $("#clinics"), true);
+                            setTimeout(function() { Consent.setDefaultConsent(userId, parentOrg);}, 500);
+                        }
+                    }
+                } else {
+                    self.updateOrgs(userId, $("#clinics"),true);
+                    var thisElement = $(this);
+                    setTimeout(function() {
+                        Consent.setConsentBySelectedOrg(userId, thisElement, isConsentWithTopLevelOrg);
+                    }, 500);
+                }
+            });
+        });
+    };
+    OrgTool.prototype.setOrgsVis = function(data, callback) {
+        callback = callback || function() {};
+        if (!data || ! data.careProvider) { callback(); return false;}
+        for (var i = 0; i < data.careProvider.length; i++) {
+            var val = data.careProvider[i];
+            var orgID = val.reference.split("/").pop();
+            if (parseInt(orgID) === 0) {
+                $("#userOrgs #noOrgs").prop("checked", true);
+                if ($("#stateSelector").length > 0) {
+                    $("#stateSelector").find("option[value='none']").prop("selected", true).val("none");
+                }
+            } else {
+                var ckOrg = $("#userOrgs input.clinic[value=" + orgID + "]");
+                if ($(".state-container").length) {
+                    if (ckOrg.length) {
+                        ckOrg.prop("checked", true);
+                        var state = ckOrg.attr("state");
+                        if (state) {
+                            $("#stateSelector").find("option[value='" + state + "']").prop("selected", true).val(i18next.t(state));
+                        }
+                        $("#clinics .state-selector-container").show();
+                        $("#stateSelector").trigger("change");
+                    }
+                    $(".noOrg-container").show();
+                } else {
+                    if (ckOrg.length) {
+                        ckOrg.prop("checked", true);
+                    } else {
+                        var topLevelOrg = $("#fillOrgs").find("legend[orgid='" + orgID + "']");
+                        if (topLevelOrg.length) {
+                            topLevelOrg.attr("data-checked", "true");
+                        }
+                    }
+                }
+            }
+        }
+        callback(data);
+    };
+    OrgTool.prototype.updateOrgs = function(userId, targetField, sync, callback) {
+        callback = callback || function() {};
+        var demoArray = {"resourceType": "Patient"}, preselectClinic = $("#preselectClinic").val();
+        if (preselectClinic) {
+            demoArray.careProvider = [{reference: "api/organization/" + preselectClinic}]; /* add this regardless of consent */
+        } else {
+            var orgIDs = $("#userOrgs input[name='organization']:checked").map(function() {
+                return {reference: "api/organization/" + $(this).val()};
+            }).get();
+            if (orgIDs && orgIDs.length > 0) {
+                demoArray.careProvider = orgIDs;
+            }
+            /**** dealing with the scenario where user can be affiliated with top level org e.g. TrueNTH Global Registry, IRONMAN, via direct database addition **/
+            $("#fillOrgs legend[data-checked]").each(function() {
+                var tOrg = $(this).attr("orgid");
+                if (tOrg) {
+                    demoArray.careProvider = demoArray.careProvider || [];
+                    demoArray.careProvider.push({reference: "api/organization/" + tOrg});
+                }
+            });
+        }
+        if ($("#aboutForm").length === 0 && (!demoArray.careProvider)) { //don't update org to none if there are top level org affiliation above
+            demoArray.careProvider = [{reference: "api/organization/" + 0}];
+        }
+        tnthAjax.putDemo(userId, demoArray, targetField, sync, function() {
+            $("#clinics").trigger("updated");
+            callback();
+        });
     };
     OrgTool.prototype.getShortName = function(orgId) {
         var shortName = "";
