+++ portal/static/js/src/modules/Utility.js
@@ -352,3 +352,5 @@ var Utility = (function() {
 export default Utility;
 export var getExportFileName = Utility.getExportFileName; /* expose common functions */
 export var getUrlParameter= Utility.getUrlParameter;
+
+
